requirements.txt
robotframework
fastapi
uvicorn
robotframework-requests